+++ examples/browser-ts-react/typings/custom.d.ts
@@ -1,1 +1,6 @@
 declare function require(module: string): any;
+
+/**
+ * This will be used for the custom markdown babel plugin.
+ */
+declare const markdown: any;